# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List all profiles
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfileRecord"
    post:
      operationId: createProfile
      tags: [profiles]
      summary: Create a new profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "201":
          description: Created profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileRecord"
  /profiles/{id}:
    delete:
      operationId: deleteProfile
      tags: [profiles]
      summary: Delete a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateProfile
      tags: [profiles]
      summary: Update profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /skin-inventory:
    get:
      operationId: listSkinInventory
      tags: [skinInventory]
      summary: Get current user skin inventory
      responses:
        "200":
          description: Skin inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkinInventoryRecord"
    post:
      operationId: giveSkin
      tags: [skinInventory]
      summary: Give a skin to a user (moderator)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiveSkinInput"
      responses:
        "201":
          description: Skin given
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkinInventoryRecord"
  /skin-inventory/user/{userId}:
    get:
      operationId: listSkinInventoryForUser
      tags: [skinInventory]
      summary: Get skin inventory for a specific user (mod panel)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Skin inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkinInventoryRecord"
  /skin-inventory/{id}:
    delete:
      operationId: removeSkin
      tags: [skinInventory]
      summary: Remove a skin from inventory (moderator)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /custom-borders:
    get:
      operationId: listCustomBorders
      tags: [customBorders]
      summary: List all custom ID borders
      responses:
        "200":
          description: List of borders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomBorderRecord"
    post:
      operationId: createCustomBorder
      tags: [customBorders]
      summary: Create a custom ID border (moderator)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomBorderInput"
      responses:
        "201":
          description: Created border
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomBorderRecord"
  /custom-borders/{id}:
    delete:
      operationId: deleteCustomBorder
      tags: [customBorders]
      summary: Delete a custom border (moderator)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /border-inventory:
    get:
      operationId: listBorderInventory
      tags: [borderInventory]
      summary: Get current user border inventory
      responses:
        "200":
          description: Border inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BorderInventoryRecord"
    post:
      operationId: giveBorder
      tags: [borderInventory]
      summary: Give a border to a user (moderator)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiveBorderInput"
      responses:
        "201":
          description: Border given
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BorderInventoryRecord"
  /border-inventory/user/{userId}:
    get:
      operationId: listBorderInventoryForUser
      tags: [borderInventory]
      summary: Get border inventory for a specific user (mod panel)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Border inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BorderInventoryRecord"
  /border-inventory/{id}:
    delete:
      operationId: removeBorder
      tags: [borderInventory]
      summary: Remove a border from inventory (moderator)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ProfileRecord:
      type: object
      required: [id, username, displayName, favoriteColor, bio, imageData, ageGroup, badges, traits, skin, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        username:
          type: string
        displayName:
          type: string
        favoriteColor:
          type: string
        bio:
          type: string
        imageData:
          type: string
        ageGroup:
          type: string
        badges:
          type: array
          items:
            type: string
        traits:
          type: array
          items:
            type: string
        skin:
          type: string
        banner:
          type: ["string", "null"]
        boarder:
          type: ["string", "null"]
        createdAt:
          type: string
    ProfileInput:
      type: object
      required: [username, displayName, favoriteColor, bio, imageData, ageGroup]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 30
        displayName:
          type: string
          minLength: 1
          maxLength: 30
        favoriteColor:
          type: string
        bio:
          type: string
          maxLength: 75
        imageData:
          type: string
        ageGroup:
          type: string
        banner:
          type: string
        traits:
          type: array
          items:
            type: string
    ProfileUpdate:
      type: object
      properties:
        skin:
          type: string
        badges:
          type: array
          items:
            type: string
        banner:
          type: ["string", "null"]
        boarder:
          type: ["string", "null"]
        username:
          type: string
        displayName:
          type: string
        favoriteColor:
          type: string
        bio:
          type: string
        imageData:
          type: string
        ageGroup:
          type: string
        traits:
          type: array
          items:
            type: string
...
    SkinInventoryRecord:
      type: object
      required: [id, userId, skinName, grantedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        skinName:
          type: string
        grantedAt:
          type: string
    GiveSkinInput:
      type: object
      required: [userId, skinName]
      properties:
        userId:
          type: integer
        skinName:
          type: string
        accessCode:
          type: string
    CustomBorderRecord:
      type: object
      required: [id, name, imageData, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        imageData:
          type: string
        createdAt:
          type: string
    CustomBorderInput:
      type: object
      required: [name, imageData]
      properties:
        name:
          type: string
          minLength: 1
        imageData:
          type: string
    BorderInventoryRecord:
      type: object
      required: [id, userId, borderName, grantedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        borderName:
          type: string
        grantedAt:
          type: string
    GiveBorderInput:
      type: object
      required: [userId, borderName]
      properties:
        userId:
          type: integer
        borderName:
          type: string
        accessCode:
          type: string