# lib/api-zod/src/generated/types/customBorderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomBorderInput {
  /** @minLength 1 */
  name: string;
  imageData: string;
}