# lib/api-zod/src/generated/types/customBorderRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomBorderRecord {
  id: number;
  name: string;
  imageData: string;
  createdAt: string;
}