# lib/api-zod/src/generated/types/giveSkinInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GiveSkinInput {
  userId: number;
  skinName: string;
  accessCode?: string;
}