# lib/api-zod/src/generated/types/profileRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileRecord {
  id: number;
  /** @nullable */
  userId?: number | null;
  username: string;
  displayName: string;
  favoriteColor: string;
  bio: string;
  imageData: string;
  ageGroup: string;
  badges: string[];
  traits: string[];
  skin: string;
  /** @nullable */
  banner?: string | null;
  /** @nullable */
  boarder?: string | null;
  createdAt: string;
}